requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0

# Streaming and Messaging
kafka-python>=2.0.2
pyspark[sql]>=3.5.0

# Vector Database and Embeddings
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.0

# Web Framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
httpx>=0.25.0

# Caching and Storage
redis>=5.0.0
psutil>=5.9.0

# Data Processing
tqdm>=4.66.0
python-dateutil>=2.8.2
pytz>=2023.3

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0

# Monitoring and Logging
prometheus-client>=0.18.0
python-json-logger>=2.0.7

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Optional: Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
